refactor: `formatDuration` を utils に切り出し
