+++ workspaces/server/src/infrastructure/repository/discord/calendar.ts
@@ -2,6 +2,7 @@ import type {
 	ICalendarEventNotify,
 	ICalendarNotifier,
 } from "../../../repository/calendar";
+import { formatDuration } from "../../../utils/date";
 
 export class DiscordCalendarNotifier implements ICalendarNotifier {
 	private webhookUrl: string;
@@ -9,55 +10,6 @@ export class DiscordCalendarNotifier implements ICalendarNotifier {
 
 	constructor(webhookUrl: string) {
 		this.webhookUrl = webhookUrl;
-	}
-
-	// 期間を表示する形式に変換する
-	// タイムゾーンについては Asia/Tokyo を明示的に指定
-	private formatDateTime(startAt: Date, endAt: Date): string {
-		if (
-			startAt.getFullYear() === endAt.getFullYear() &&
-			startAt.getMonth() === endAt.getMonth() &&
-			startAt.getDate() === endAt.getDate()
-		) {
-			const date = startAt.toLocaleDateString("ja-JP", {
-				month: "2-digit",
-				day: "2-digit",
-				weekday: "short",
-				timeZone: "Asia/Tokyo",
-			});
-			const startTimestamp = startAt.toLocaleTimeString("ja-JP", {
-				hour: "2-digit",
-				minute: "2-digit",
-				timeZone: "Asia/Tokyo",
-			});
-			const endTimestamp = endAt.toLocaleTimeString("ja-JP", {
-				hour: "2-digit",
-				minute: "2-digit",
-				timeZone: "Asia/Tokyo",
-			});
-
-			return `${date} ${startTimestamp} - ${endTimestamp}`;
-		}
-
-		const startAtDate = startAt.toLocaleDateString("ja-JP", {
-			month: "2-digit",
-			day: "2-digit",
-			weekday: "short",
-			hour: "2-digit",
-			minute: "2-digit",
-			timeZone: "Asia/Tokyo",
-		});
-
-		const endAtDate = endAt.toLocaleDateString("ja-JP", {
-			month: "2-digit",
-			day: "2-digit",
-			weekday: "short",
-			hour: "2-digit",
-			minute: "2-digit",
-			timeZone: "Asia/Tokyo",
-		});
-
-		return `${startAtDate} - ${endAtDate}`;
 	}
 
 	private embedBuilder(embedInfo: ICalendarEventNotify & { color: number }) {
@@ -68,7 +20,7 @@ export class DiscordCalendarNotifier implements ICalendarNotifier {
 			fields: [
 				{
 					name: "日時",
-					value: this.formatDateTime(embedInfo.startAt, embedInfo.endAt),
+					value: formatDuration(embedInfo.startAt, embedInfo.endAt),
 				},
 				{
 					name: "場所",
